+++ src/web/RestApi.h
@@ -7,11 +7,6 @@
 #define __WEB_API_H__
 
 #include "../utils/dbg.h"
-#ifdef ESP32
-#include "AsyncTCP.h"
-#else
-#include "ESPAsyncTCP.h"
-#endif
 #include "../appInterface.h"
 #include "../hm/hmSystem.h"
 #include "../utils/helper.h"